requirements.txt
Django==5.1
biopython==1.84
django-ninja==1.3.0
pandas==2.2.2
numpy==2.0.1
joblib==1.4.2
seaborn==0.13.2
requests==2.32.3
plotly==5.23.0
django-filter==24.3
django-matomo==0.1.6
gunicorn==20.1.0
whitenoise==6.4.0
django-debug-toolbar==4.4.6